Added modularity by adding routers in diffrent folder
